Added validation and logging

+++ src/post/post.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 
@@ -23,16 +23,34 @@ export class PostService {
     }
     
     async show(id: string){
-        return await this.postRepository.findOne({ where: { id } });
+        const post = await this.postRepository.findOne({ where: { id } });
+
+        if(!post){
+            throw new HttpException("Not Found", HttpStatus.NOT_FOUND);
+        }
+
+        return post;
     }
 
     async update(id: string, data: Partial<PostDTO>){
+        let post = await this.postRepository.findOne( { where: { id } } );
+        
+        if(!post){
+            throw new HttpException("Not Found", HttpStatus.NOT_FOUND);
+        }
         await this.postRepository.update({ id }, data);
-        return await this.postRepository.findOne( { id } );
+        post = await this.postRepository.findOne({ where: { id } });
+
+        return post;
     }
 
     async delete(id: string){
+        const post = await this.postRepository.findOne( { where: { id } } );
+        
+        if(!post){
+            throw new HttpException("Not Found", HttpStatus.NOT_FOUND);
+        }
         await this.postRepository.delete({ id });
-        return { deleted : true };
+        return post;
     }
 }
